pyproject.toml
[project]
name = "clinical-trial-matcher"
version = "0.1.0"
description = "Multi-agent system that matches patients to recruiting clinical trials using LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4",
    "langchain>=0.3",
    "langgraph-checkpoint-postgres>=2.0",
    "langsmith>=0.3",
    "langchain-anthropic>=0.3",
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "tenacity>=9.0",
    "psycopg[binary]>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
